requirements.txt
numpy==1.21.5
numpy==1.21.6
onnx==1.14.1
onnxruntime_gpu==1.10.0
opencv_python==4.5.5.64
Requests==2.32.3
torch==1.8.0+cu111
torchvision==0.9.0+cu111
